requirements.txt
black==23.1.0
flake8==6.0.0
opencv-python==4.7.0.72
pillow==9.4.0
pre-commit==3.1.1
pytest==7.2.1
coverage==7.2.1
pytest_cov==4.0.0
pytest-mock==3.10.0
pyyaml==6.0
wandb==0.14.0
jupyter==1.0.0
nvidia-tao